backend/requirements.txt
pipecat-ai[deepgram,groq]>=0.0.53
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
